+++ lib/themes.dart
@@ -111,7 +111,7 @@ class Themes {
       acceptLight: const Color(0xFFBFE3B9),
       acceptLighter: const Color(0xFFD9FDD3),
       acceptLightest: const Color(0xFFF2FDED),
-      decline: const Color(0x7FFF0000),
+      decline: const Color(0xFFFF0000),
       danger: const Color(0xFFF44336),
       warning: const Color(0xFFFF9800),
       userColors: const [
@@ -686,6 +686,8 @@ class Palette {
     required this.acceptLighter,
     required this.acceptLightest,
     required this.decline,
+    Color? declineOpacity50,
+    Color? declineOpacity88,
     required this.danger,
     required this.warning,
     required this.userColors,
@@ -725,7 +727,9 @@ class Palette {
         onBackgroundOpacity50 =
             onBackgroundOpacity50 ?? onBackground.withOpacity(0.50),
         onBackgroundOpacity70 =
-            onBackgroundOpacity70 ?? onBackground.withOpacity(0.70);
+            onBackgroundOpacity70 ?? onBackground.withOpacity(0.70),
+        declineOpacity50 = declineOpacity50 ?? decline.withOpacity(0.50),
+        declineOpacity88 = declineOpacity88 ?? decline.withOpacity(0.88);
 
   /// Primary [Color] of the application.
   ///
@@ -960,9 +964,17 @@ class Palette {
 
   /// Indicator of rejection or cancellation in various elements of the user
   /// interface.
+  final Color decline;
+
+  /// 88% opacity of the [decline] color.
   ///
   /// Used in decline call button.
-  final Color decline;
+  final Color declineOpacity88;
+
+  /// 50% opacity of the [decline] color.
+  ///
+  /// Used in decline call button.
+  final Color declineOpacity50;
 
   /// [Color] used to indicate dangerous or critical elements in the user
   /// interface.
@@ -1073,6 +1085,10 @@ class Palette {
       acceptLightest:
           Color.lerp(color.acceptLightest, other.acceptLightest, t)!,
       decline: Color.lerp(color.decline, other.decline, t)!,
+      declineOpacity50:
+          Color.lerp(color.declineOpacity50, other.declineOpacity50, t)!,
+      declineOpacity88:
+          Color.lerp(color.declineOpacity88, other.declineOpacity88, t)!,
       danger: Color.lerp(color.danger, other.danger, t)!,
       warning: Color.lerp(color.warning, other.warning, t)!,
       userColors: