+++ CHANGELOG.md
@@ -6,6 +6,23 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
+## [0.1.0-alpha.13] · 2024-??-??
+[0.1.0-alpha.13]: /../../tree/v0.1.0-alpha.13
+
+[Diff](/../../compare/v0.1.0-alpha.12.1...v0.1.0-alpha.13) | [Milestone](/../../milestone/15)
+
+### Changed
+
+- UI:
+    - Media panel:
+        - Redesigned incoming/outgoing call. ([#832], [#812])
+
+[#812]: /../../issues/812
+[#832]: /../../pull/832
+
+
+
+
 ## [0.1.0-alpha.12.1] · 2024-02-08
 [0.1.0-alpha.12.1]: /../../tree/v0.1.0-alpha.12.1